See: https://github.com/neovim/neovim/pull/32644

Functional tests call `clear()` before each test, which spawns a new neovim instance, and start using it immediately after. This forces each test to wait until the newly spawned neovim is ready.

Added a queue of 10 neovim instances that are spawned before they are needed. When `clear()` is called, the oldest instance in the queue is used. This makes it more likely that the instance is ready by the time it is needed. Queue is refilled when instances are removed from it.

This reduces total time taken by functional tests.

### Additional info

RPC server name is set at runtime, so some entries in `.nvimlog` don't have the server name. However, since pid of the process is printed, it's easy to know which message belongs to which session.

Added `clear_without_pool()` function for the cases where all log entries need to have the same name, e.g. `log_spec.lua`. Also needed for `diff_spec.lua` on windows, see previous PR.

### Results

I updated the results. Does anyone know why the tests became so much slower since February 2025? It's almost 1.5x slower.

| | current | new |
|-|-|-|
| ubuntu asan clang functionaltest | 19m 46s | 16m 36s |
| ubuntu tsan clang functionaltest | 19m 14s | 19m 6s |
| ubuntu release gcc functionaltest | 10m 23s | 9m 11s |
| macos intel clang functionaltest | 11m 15s | 10m 38s |
| macos arm clang functionaltest | 8m 9s | 7m 18s |
| ubuntu puc-lua gcc functionaltest | 9m 5s | 8m 56s |
| windows / windows (functional) | 15m 12s | 11m 4s |

<details>

current:
https://github.com/neovim/neovim/actions/runs/16103434692/job/45435514123

new:
https://github.com/vanaigr/neovim/actions/runs/16103544445/job/45435754589

</details>