@justinmk 2025-07-07
> Does anyone know why the tests became so much slower since February 2025? It's almost 1.5x slower.

Comparing the time cost of the `ubuntu asan clang functionaltest` job:

- for [current master](https://github.com/neovim/neovim/actions/runs/16104186313/job/45437479137), `8134 tests` run in 15:54
- for [current 0.11 branch](https://github.com/neovim/neovim/pull/34812), `7972 tests` run in 15:22
- a random commit from dec 2024 the *total* time is 16 minutes, but can't inspect the job details :( there were probably 100-200 fewer tests then.

we also enabled more tests on Windows, in master.