@vanaigr 2025-07-07
Noticed that `exec_lua` in `lsp_spec.lua` references `lsp` which is not always defined, but no tests fail:

https://github.com/neovim/neovim/blob/5973328edaedb04425e2f6cb63b2efab4aeb5fcb/test/functional/plugin/lsp_spec.lua#L79-L83

But if I add a print statement:
```diff
diff --git a/test/functional/plugin/lsp_spec.lua b/test/functional/plugin/lsp_spec.lua
index 15bb6e2006..7b1d753293 100644
--- a/test/functional/plugin/lsp_spec.lua
+++ b/test/functional/plugin/lsp_spec.lua
@@ -78,6 +78,7 @@ describe('LSP', function()
 
   after_each(function()
     stop()
+    print(exec_lua('vim.inspect(lsp)'))
     exec_lua('lsp.stop_client(lsp.get_clients(), true)')
     api.nvim_exec_autocmds('VimLeavePre', { modeline = false })
   end)
```

then it prints `nil`, and there's a lot of errors, and some of the tests fail.